backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
python-multipart==0.0.12
google-generativeai==0.8.3
Pillow==10.4.0

# Testes
pytest==8.3.3
httpx==0.27.2
